requirements.txt
Flask==2.3.2
requests==2.31.0
pandas
python-dotenv==1.0.0
textblob==0.17.1
mysql-connector-python==8.2.0